Add files via upload
Save and create new directory are commented out, and only PC x1 y2 are computed; this is for ease of testing. The cancer type list is alphabetised for ease of access on the part of the user. The list indexing was uncommented with no issue since the indexing error was down to incorrectly referencing the list. Returns out_plots, then out_plots are saved as a list in the app.R, and the list referenced and graph plotted from here. Current_gene_names changed to current_gene because the names() function was not needed and was causing the figure label to be blank.