Add files via upload
RadioButtons are now checkboxGroupInput so that the user can select multiple cancer types. The user now automatically views the graph which equates to the first of each of their variables. The cancer type list was updated since some of them did not have rds files (it is assumed that this is because they did not have enough RNA-seq data samples after being filtered); this was causing an issue when the user selected the cancer type that did not have an rds file.