Add files via upload
The display options for cancer type are now the name instead of the code (on the input drop-down menus and the graph). The graphs now have a title, and the label for gene expression is simply transcripts per million. Plot images are formatted to a reasonable aspect ratio as standard for download.